refactor model
