--- scripts/benchmark.py
@@ -1,8 +1,0 @@
-"""
-Benchmark module for YOLO-One
-
-TODO: Implement benchmark.py functionality
-"""
-
-# TODO: Add imports
-# TODO: Add implementation
