--- scripts/download_models.py
@@ -1,8 +1,0 @@
-"""
-Download_Models module for YOLO-One
-
-TODO: Implement download_models.py functionality
-"""
-
-# TODO: Add imports
-# TODO: Add implementation
