--- scripts/eval.py
@@ -1,8 +1,0 @@
-"""
-Eval module for YOLO-One
-
-TODO: Implement eval.py functionality
-"""
-
-# TODO: Add imports
-# TODO: Add implementation
