--- scripts/tools/analyze_model.py
@@ -1,8 +1,0 @@
-"""
-Analyze_Model module for YOLO-One
-
-TODO: Implement analyze_model.py functionality
-"""
-
-# TODO: Add imports
-# TODO: Add implementation
