--- scripts/tools/convert_dataset.py
@@ -1,8 +1,0 @@
-"""
-Convert_Dataset module for YOLO-One
-
-TODO: Implement convert_dataset.py functionality
-"""
-
-# TODO: Add imports
-# TODO: Add implementation
