--- scripts/tools/profile_performance.py
@@ -1,8 +1,0 @@
-"""
-Profile_Performance module for YOLO-One
-
-TODO: Implement profile_performance.py functionality
-"""
-
-# TODO: Add imports
-# TODO: Add implementation
