--- scripts/tools/validate_installation.py
@@ -1,8 +1,0 @@
-"""
-Validate_Installation module for YOLO-One
-
-TODO: Implement validate_installation.py functionality
-"""
-
-# TODO: Add imports
-# TODO: Add implementation
