--- scripts/train.py
@@ -1,8 +1,0 @@
-"""
-Train module for YOLO-One
-
-TODO: Implement train.py functionality
-"""
-
-# TODO: Add imports
-# TODO: Add implementation
