+++ yolo_one/models/yolo_one_backbone_box.py
@@ -1,0 +1,12 @@
+"""
+Iatrax Team - 2025 - https://iatrax.com
+
+LICENSE: MIT
+
+BACKBONE MODULE FOR YOLO-ONE
+
+This module contains the backbone for YOLO-ONE, which is a single-class object detector.
+"""
+import torch
+import torch.nn as nn
+from typing import List, Dict, Any
